Move routes to the emission pathways page instead of nesting them
