+++ app/javascript/app/pages/emission-pathways/emission-pathways-component.jsx
@@ -4,6 +4,7 @@ import Header from 'components/header';
 import Intro from 'components/intro';
 import AnchorNav from 'components/anchor-nav';
 import Sticky from 'react-stickynode';
+import { renderRoutes } from 'react-router-config';
 
 import anchorNavRegularTheme from 'styles/themes/anchor-nav/anchor-nav-regular.scss';
 import layout from 'styles/layout.scss';
@@ -12,7 +13,7 @@ import styles from './emission-pathways-styles.scss';
 class EmissionPathways extends PureComponent {
   // eslint-disable-line react/prefer-stateless-function
   render() {
-    const { route, anchorLinks } = this.props;
+    const { route, anchorLinks, routeLinks } = this.props;
     return (
       <div>
         <Header route={route}>
@@ -32,9 +33,10 @@ class EmissionPathways extends PureComponent {
           route.sections.map(section => (
             <div key={section.hash} className={styles.section}>
               <div id={section.hash} className={styles.sectionHash} />
-              <section.component routes={section.routes} />
+              <section.component routeLinks={routeLinks} />
             </div>
           ))}
+        {renderRoutes(route.routes)}
       </div>
     );
   }
@@ -42,7 +44,8 @@ class EmissionPathways extends PureComponent {
 
 EmissionPathways.propTypes = {
   route: PropTypes.object.isRequired,
-  anchorLinks: PropTypes.array.isRequired
+  anchorLinks: PropTypes.array.isRequired,
+  routeLinks: PropTypes.array.isRequired
 };
 
 export default EmissionPathways;
