+++ app/javascript/app/pages/emission-pathways/emission-pathways-selectors.js
@@ -1,8 +1,10 @@
 import { createSelector } from 'reselect';
 
-const getSections = data => data.route.sections || null;
-export const getRoutes = data => data.route.routes || null;
-export const getSearch = data => data.location.search || null;
+const getSections = routeData => routeData.route.sections || null;
+const getSearch = routeData => routeData.location.search || null;
+const getHash = routeData => routeData.hash || null;
+const getRoutes = routeData => routeData.route.routes || null;
+
 export const getAnchorLinks = createSelector(
   [getSections, getSearch],
   (sections, search) =>
@@ -14,7 +16,19 @@ export const getAnchorLinks = createSelector(
     }))
 );
 
+export const getRouteLinks = createSelector(
+  [getRoutes, getHash, getSearch],
+  (routes, hash, search) =>
+    routes &&
+    routes.filter(r => r.anchor).map(route => ({
+      label: route.label,
+      path: route.path,
+      search,
+      hash
+    }))
+);
+
 export default {
   getAnchorLinks,
-  getRoutes
+  getRouteLinks
 };
