+++ app/javascript/app/pages/emission-pathways/emission-pathways.js
@@ -1,11 +1,15 @@
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
-import { getAnchorLinks } from './emission-pathways-selectors';
+import { getAnchorLinks, getRouteLinks } from './emission-pathways-selectors';
 import Component from './emission-pathways-component';
 
-const mapStateToProps = (state, { route, location }) => ({
-  query: location.search,
-  anchorLinks: getAnchorLinks({ route, location })
-});
+const mapStateToProps = (state, { route, location }) => {
+  const routeData = { route, location, hash: location.hash };
+  return {
+    query: location.search,
+    anchorLinks: getAnchorLinks(routeData),
+    routeLinks: getRouteLinks(routeData)
+  };
+};
 
 export default withRouter(connect(mapStateToProps, null)(Component));
